Parse options correctly
Fix #1
Options have already passed through `nopt` module in the build file. Nopt successfully parses all options which results in `argv.remain` being an empty array.
This ultimately leads to a bug in the `parseOpts` function where the options undergo a second parsing from nopt, whereas they are actually ready to be consumed.

Fix #2
parseOpts function referenced `this.root` but the `this` object is not the `new Api` as expected. This leads to an exception upon calling `path.relative(this.root, ...)`.

# Conflicts:
#	bin/templates/cordova/lib/build.js